Refactor I2C control and test logic for Rogallo
Changed sendCtrl to accept a float instead of int for asymmetric deflection control. Added MOTOR_PACKET_LEN constant and improved documentation. Commented out previous main test logic and replaced with a new threaded I2C slave handler and test loop in main.cpp for MCPS simulation.

+++ GNC/ctrldRogallo.h
@@ -97,7 +97,7 @@ class ctrldRogallo {
         float computeCtrl(float heading_error, float dt); // output in [-1, 1]
 
         // MC/PS comms
-        uint8_t sendCtrl(int ctrl);
+        uint8_t sendCtrl(float ctrl);
         char* requestMotorPacket(void);
 
         uint32_t apogeeDetection(double prevAlt, double currAlt);
